Add support for granting or abstaining on transient objects
Add a strategy to test an object transientness and subsequently either
grant or abstain. If no strategy is set, then default to the parent
class behaviour of raising an IllegalArgumentException.

Because we need to modify AclEntryVoter's default behavior, we minimally
subclass it and use it in the configuration.

Use Spring Security-provided interface for AclService in
AclEntryCollectionVoter.